Login html and js, moved everything related to login into authentication.go
