+++ views/login/login.html
@@ -1,0 +1,64 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <title>Caraway Volunteer Management System</title>
+        <meta charset="UTF-8">
+        <!-- Latest compiled and minified CSS -->       
+        <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
+
+        <!-- jQuery library -->
+        <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
+
+        <!-- Latest compiled JavaScript -->
+        <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
+        <link href='https://fonts.googleapis.com/css?family=Sofia' rel='stylesheet'>
+        <meta name="viewport" content="width=device-width, initial-scale=1">
+    </head>
+    <body>
+        <header>
+            <div class="col-sm-9 col-sm-offset-1">
+                <h2 style="font-family: 'Sofia';" >Caraway Volunteer Management System</h2>
+            </div>
+            
+        </header>
+        <br><br><br><br>
+        <main>
+
+                <div class="container" style="width: 70%" >
+                    <div class ="well"> 
+                       <div class="form-group">
+                            <label for="usr">Username:</label>
+                            <input type="text" class="form-control text-left" id="usr">
+                          </div>
+                          <div class="form-group">
+                            <label for="pwd">Password:</label>
+                            <input type="password" class="form-control text-left" id="pwd">
+                          </div>           
+                          <div class="form-group text-right">
+                            <button type="button" class="btn btn-primary" id="loginbtn">Login</button>
+                          </div>          
+                    </div>
+
+                </div>
+    
+
+        </main>
+        <footer> 
+            <div class="navbar navbar-fixed-bottom" style="background-color: black;height: 20px">
+                <div class="row">
+                        <div class="col-lg-12">
+                            <div class="container" style="color: white">
+                                <p>
+                                    this is a footer
+                                </p>
+                                <p>
+                                        &copy; Team gold CMPT 395
+                                </p>
+                            </div>
+                        </div>
+                </div>
+            </div>
+        </footer>
+        <script src="login.js"></script>
+    </body>
+</html>
